requirements.txt
numpy==1.14.3
Django==1.11
django-jasmine==0.5.1
django-simple-captcha==0.5.6
django-simple-email-confirmation==0.23
Pillow==5.1.0
six==1.11.0
tensorflow==1.8.0
keras==2.1.6
annoy==1.12.0
h5py==2.7.1
tqdm==4.23.3